split external session importer into focused modules
