+++ codex-rs/app-server/src/external_agent_sessions/import.rs
@@ -1,0 +1,220 @@
+use super::ConversationMessage;
+use super::ImportedExternalAgentSession;
+use super::MessageRole;
+use super::records::conversation_messages;
+use super::records::custom_title_from_records;
+use super::records::project_root_from_records;
+use super::records::read_records;
+use super::summarize_for_label;
+use codex_protocol::models::ContentItem;
+use codex_protocol::models::ResponseItem;
+use codex_protocol::protocol::AgentMessageEvent;
+use codex_protocol::protocol::EventMsg;
+use codex_protocol::protocol::RolloutItem;
+use codex_protocol::protocol::TurnCompleteEvent;
+use codex_protocol::protocol::TurnStartedEvent;
+use codex_protocol::protocol::UserMessageEvent;
+use std::io;
+use std::path::Path;
+
+pub(crate) fn load_session_for_import(
+    path: &Path,
+) -> io::Result<Option<ImportedExternalAgentSession>> {
+    let records = read_records(path)?;
+    let Some(cwd) = project_root_from_records(&records) else {
+        return Ok(None);
+    };
+    let messages = conversation_messages(&records);
+    let rollout_items = rollout_items_from_messages(&messages);
+    if rollout_items.is_empty() {
+        return Ok(None);
+    }
+    let title = custom_title_from_records(&records).or_else(|| {
+        messages
+            .iter()
+            .find(|message| message.role == MessageRole::User)
+            .map(|message| summarize_for_label(&message.text))
+    });
+    Ok(Some(ImportedExternalAgentSession {
+        cwd,
+        title,
+        rollout_items,
+    }))
+}
+
+fn rollout_items_from_messages(messages: &[ConversationMessage]) -> Vec<RolloutItem> {
+    let mut items = Vec::new();
+    let mut current_turn: Option<(String, Option<String>)> = None;
+    let mut user_turn_count = 0usize;
+
+    for message in messages {
+        match message.role {
+            MessageRole::User => {
+                if let Some((turn_id, last_agent_message)) = current_turn.take() {
+                    items.push(turn_complete_item(turn_id, last_agent_message, None));
+                }
+                user_turn_count += 1;
+                let turn_id = format!("external-import-turn-{user_turn_count}");
+                items.push(RolloutItem::EventMsg(EventMsg::TurnStarted(
+                    TurnStartedEvent {
+                        turn_id: turn_id.clone(),
+                        started_at: message.timestamp,
+                        model_context_window: None,
+                        collaboration_mode_kind: Default::default(),
+                    },
+                )));
+                items.push(response_item(message));
+                items.push(RolloutItem::EventMsg(EventMsg::UserMessage(
+                    UserMessageEvent {
+                        message: message.text.clone(),
+                        images: None,
+                        local_images: Vec::new(),
+                        text_elements: Vec::new(),
+                    },
+                )));
+                current_turn = Some((turn_id, None));
+            }
+            MessageRole::Assistant => {
+                let Some((_, last_agent_message)) = current_turn.as_mut() else {
+                    continue;
+                };
+                items.push(response_item(message));
+                items.push(RolloutItem::EventMsg(EventMsg::AgentMessage(
+                    AgentMessageEvent {
+                        message: message.text.clone(),
+                        phase: None,
+                        memory_citation: None,
+                    },
+                )));
+                *last_agent_message = Some(message.text.clone());
+            }
+        }
+    }
+
+    if let Some((turn_id, last_agent_message)) = current_turn {
+        let completed_at = messages.last().and_then(|message| message.timestamp);
+        items.push(turn_complete_item(
+            turn_id,
+            last_agent_message,
+            completed_at,
+        ));
+    }
+
+    items
+}
+
+fn response_item(message: &ConversationMessage) -> RolloutItem {
+    let content = match message.role {
+        MessageRole::Assistant => ContentItem::OutputText {
+            text: message.text.clone(),
+        },
+        MessageRole::User => ContentItem::InputText {
+            text: message.text.clone(),
+        },
+    };
+    RolloutItem::ResponseItem(ResponseItem::Message {
+        id: None,
+        role: match message.role {
+            MessageRole::Assistant => "assistant".to_string(),
+            MessageRole::User => "user".to_string(),
+        },
+        content: vec![content],
+        phase: None,
+    })
+}
+
+fn turn_complete_item(
+    turn_id: String,
+    last_agent_message: Option<String>,
+    completed_at: Option<i64>,
+) -> RolloutItem {
+    RolloutItem::EventMsg(EventMsg::TurnComplete(TurnCompleteEvent {
+        turn_id,
+        last_agent_message,
+        completed_at,
+        duration_ms: None,
+        time_to_first_token_ms: None,
+    }))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use codex_app_server_protocol::build_turns_from_rollout_items;
+    use serde_json::Value as JsonValue;
+    use std::path::Path;
+    use tempfile::TempDir;
+
+    #[test]
+    fn builds_visible_turns_for_imported_history() {
+        let root = TempDir::new().expect("tempdir");
+        let project_root = root.path().join("repo");
+        std::fs::create_dir_all(&project_root).expect("project root");
+        let path = root.path().join("session.jsonl");
+        std::fs::write(
+            &path,
+            jsonl(&[
+                record("user", "first request", &project_root),
+                record("assistant", "first answer", &project_root),
+                record("user", "second request", &project_root),
+            ]),
+        )
+        .expect("session");
+
+        let imported = load_session_for_import(&path)
+            .expect("load")
+            .expect("session");
+        let turns = build_turns_from_rollout_items(&imported.rollout_items);
+
+        assert_eq!(turns.len(), 2);
+        assert_eq!(turns[0].items.len(), 2);
+        assert_eq!(turns[1].items.len(), 1);
+    }
+
+    #[test]
+    fn loads_custom_title_for_imported_session() {
+        let root = TempDir::new().expect("tempdir");
+        let project_root = root.path().join("repo");
+        std::fs::create_dir_all(&project_root).expect("project root");
+        let path = root.path().join("session.jsonl");
+        std::fs::write(
+            &path,
+            jsonl(&[
+                record("user", "first request", &project_root),
+                custom_title_record("named by source app"),
+            ]),
+        )
+        .expect("session");
+
+        let imported = load_session_for_import(&path)
+            .expect("load")
+            .expect("session");
+
+        assert_eq!(imported.title.as_deref(), Some("named by source app"));
+    }
+
+    fn record(role: &str, text: &str, cwd: &Path) -> JsonValue {
+        let timestamp = chrono::Utc::now().to_rfc3339_opts(chrono::SecondsFormat::Secs, true);
+        serde_json::json!({
+            "type": role,
+            "cwd": cwd,
+            "timestamp": timestamp,
+            "message": { "content": text }
+        })
+    }
+
+    fn custom_title_record(title: &str) -> JsonValue {
+        serde_json::json!({
+            "type": "custom-title",
+            "customTitle": title,
+        })
+    }
+
+    fn jsonl(records: &[JsonValue]) -> String {
+        records
+            .iter()
+            .map(JsonValue::to_string)
+            .collect::<Vec<_>>()
+            .join("\n")
+    }
+}
